AWS_Sagemaker_Portfolio/Sagemaker-Freelance-Projects/Cifar10 Classification Using Sagemaker Tensorflow/src/requirements.txt
keras
tensorflow-gpu
matplotlib